Run functions for setting internal variables after mode hooks.
Also declare the configuration variables for noise macros and macros with
semicolons as safe variables (when given suitable arguments).

* cc-mode.el (c-basic-common-init): Remove the call to
c-make-noise-macro-regexps.
(c-mode, c++-mode, objc-mode): Move c-make-noise-macro-regexps and
c-make-macro-with-semi-re to after the invocation of the mode hook.

* cc-vars.el: (c-string-list-p, c-string-or-string-list-p): New functions.
(c-noise-macro-names, c-noise-macro-with-parens-names): give the
safe-local-variable property c-string-list-p.
(c-macro-names-with-semicolon): give the safe-local-variable property
c-string-or-string-list-p.

* cc-mode.texi (Macros with ;, Noise Macros): Note that it's not necessary to
call the regexp generating functions after setting the pertinent configuration
values in a mode hook.