CC Mode internal cache: Handle a cache pos being inside a two-char construct.
Cache c-state-semi-nonlit-pos-cache was failing when a cache position was,
e.g., between the two characters of an opening comment "/*", and additionally
there were an odd number of quote marks (apostrophes) in the comment.  This
happened in .../src/xdisp.c in the Emacs master branch around 2017-05-02 at
buffer position 615001.

* cc-defs.el (c-emacs-features): Repurpose symbol pps-extended-state to mean
that there are at least 11 elements in the parser state.

* cc-engine.el (c-cache-to-parse-ps-state, c-parse-ps-state-to-cache): Rewrite
these to use enhanced cache element list types which indicate potentially
being inside two-char constructs.
(c-parse-ps-state-below): Rewrite to use the new versions of the above two
functions.