Cope better with C++ and Objective-C protection keywords in class declarations
This fix fixes the fontification of method inside a class at the time it is
typed, when there is a protection keyword clause preceding it.

* cc-engine.el (c-forward-keyword-clause): Handle protection keywords.
(c-looking-at-decl-block): Avoid scanning forward over protection keyword
clauses too eagerly.

* cc-langs.el (c-protection-key c-post-protection-token): New lang defconsts
and defvars.

* cc-mode.el (c-fl-decl-start): When we encounter a protection keyword
following a semicolon or brace, move forward over it before attempting to
parse a type.