Fix another "wrong side of point" error in CC Mode.
This fixes (a follow-up to) bug #28850.
A internal generated form for scanning text to fontify had a LIMIT parameter.
It also locally bound LIMIT to a value possibly beyond the original LIMIT,
allowing point to move beyond the original LIMIT, and to create the wrong side
error.  Fix it by checking point is not beyond LIMIT in the outer context
before using it.

* cc-fonts.el (c-make-font-lock-search-form): Add a new parameter CHECK-POINT
which, when non-nil, directs the function to generate a check on point.
(c-make-font-lock-context-search-function): Invoke the above function with new
argument value t.