Correctly indent C++14 brace lists which are a second argument to a function.
In particular, don't indent contained brace lists in "staircase" fashion.
This fixes bug #28623.

* cc-engine.el (c-looking-at-or-maybe-in-bracelist): When testing for being
enclosed in parens, recognise also a brace directly following a comma, as well
as a brace being the first thing inside the paren.  Enhance the return value,
by indicating when we're directly inside an open paren.
(c-inside-bracelist-p): Add an extra argument ACCEPT-IN-PARAM which indicates
whether we will accept a bracelist directly inside an open parenthesis.
Simplify the manipulation of PAREN-STATE by dispensing with variable LIM and
using c-pull-open-brace.  Enhance the return value, respecting the new
argument.
(c-guess-basic-syntax): Save a copy of the initial parse-state in the new
variable STATE-CACHE.  Use this variable in place of C-STATE-CACHE throughout
the function.  At CASE 7B, call c-inside-bracelist-p with extra argument nil.
At CASE 9, call that function with extra argument t.

* bracelist-2.res.  Reflect the now more accurate parsing available.
* bracelist-25.{cc,res,face}.  Re-indent a brace list, and recalculate the
.res and .face files.
* bracelist-26.{cc,res}.  New test files.
* bracelist-27.{cc,res}.  New test files.