Correct the indentation of C99's compound literals.
* cc-engine.el (c-looking-at-statement-block): Amend so that if there is only
syntactic whitespace in a brace block, it is regarded as a statement block.
Also, if there is no semicolon or comma delimiter, treat as a statement block
when there is a keyword.
(c-guess-basic-syntax): CASE 9 test: Regard a brace as starting a brace block
when its contents indicate a brace block.

* bracelist-28.{c,res}: New test files.
* constants-3.face: Correct this facefile after amendments unrelated to the
above.