Don't misfontify "foo ()" inside C++ initialization parentheses as a type
Also recognize and handle function names introduced by "extern" inside a
function.

* cc-engine.el (c-forward-decl-or-cast-1): Add a new element to the result
list which is t when our declaration is, or is to be treated as, being at top
level.

* cc-fonts.el (c-get-fontification-context): Detect being inside a C++ uniform
initialization and return (not-decl nil) for this case.
(c-font-lock-declarations): Use the new element 4 of the result of
c-forward-decl-or-cast-1.

* cc-langs.el (c-make-top-level-kwds, c-make-top-level-key): New lang
consts/vars.

* statement-26.{cc,face}: New test files.

* templates-12.face: Recalculate this face file.