Detect when we hit limit in backward search in c-just-after-func-arglist-p
This is a correction of the faulty patch made on 2019-01-03.

This fixes a bug reported by Yasushi SHOJI <yasushi.shoji@gmail.com> to
emacs-devel on 2018-11-26.

* cc-engine.el (c-beginning-of-statement-1): Add new parameter HIT-LIM which,
if non-nil causes the function to return nil rather than 'same when we reach
the backward search limit without finding the beginning of statement.
(c-just-after-func-arglist-p): Supply argument t to this new parameter in call
to c-beginning-of-statement-1.