Distinguish better between nil and `same' in c-beginning-of-statement-1
This corrects a slight flaw in today's previous commits for this function.

* cc-mode (c-bos-pop-state, c-bos-pop-state-and-retry): set pre-stmt-found to
t when throwing nil to the catch form in c-beginning-of-statement-1 to
terminate its loop.
(c-beginning-of-statement-1): New local variable pre-stmt-found, which gets
set to t to indicate some backward stop to further scanning has been found.
This contrasts with the case where the search limit is passed.  Use this
variable in deciding whether or not the result of the function should be nil.