cc-mode.texi: Work around makeinfo alignment bug.  Fix problem with ss index
* cc-mode.texi (top level): Using txicommandconditionals to differentiate
between the C and perl versions of Texinfo, create an "ss index" unless we are
both using the C Texinfo and are building the .dvi output format.
(Config Basics): Work around a perl Texinfo alignment bug by writing a
separate version of an item list structure for this version, simplifying it
considerably.