Correct the indentation of CC Mode brace lists
while preserving the indentation of nested C++ uniform initialization.

* cc-align.el (c-lineup-2nd-brace-entry-in-arglist)
(c-lineup-class-decl-init-+, c-lineup-class-decl-init-after-brace): New
indentation functions.

* cc-engine.el (c-forward-class-decl): New function.
(c-do-declarators): New function, partially extracted from
c-font-lock-declarators, which now calls the new function.
(c-inside-bracelist-p): Amend the introductory comment.
(c-add-stmt-syntax): Add code to prevent the spurious recognition of a
'defun-block-intro when a brace pair is used as an initializer.
(c-evaluate-offset): No longer ignore vectors of length >= 2.
(c-calc-offset): Refactor clumsily nested `if'/`or' into a cond form.

* cc-fonts.el (c-font-lock-declarators): Replace the bulk of this function by
a call to the new c-forward-class-decl.

* cc-langs.el (c-type-decl-prefix-key): Recognize "~" as a type decl operator.

* cc-mode.el (c-fl-decl-start): While searching backward for a "}" at an EOD,
deal with actually finding the end of a brace list.

* cc-mode.texi (List Line-Up): document c-lineup-2nd-brace-entry-in-arglist,
c-lineup-class-decl-init-+, and c-lineup-class-decl-init-after-brace.

* cc-styles.el (c-style-alist): In styles "gnu", "bsd", "stroustrup",
"python", and "java", change the offset for brace-list-intro from the default
value or c-lineup-arglist-intro-after-paren to a list beginning with the
symbol first, followed by two of the new alignment functions, followed by +.

* cc-vars.el (c-offset-alist): Change the default value of brace-list-entry
from c-lineup-under-anchor back to 0.

* 000-tests.el (cc-test-teststyle): Set the entry in c-offsets-alist for
brace-list-intro to the same as in c-style-alist, above.

* bracelist-2.pike: Reindent.

* bracelist-25.{cc,res,face}: Reindent the .cc, then regenerate the .res and
.face

* bracelist-31.{c,res}: Reindent the .c, then regenerate the .res.

* bracelist-33.{cc,res}: New test files.

* bracelist-9.c: Reindent.

* class-21.c: Reindent.