Analyze C++ method with & or && ref-qualifier as defun, not brace list
Also firm up detection of beginning of brace list in
c-looking-at-or-maybe-in-bracelist.

* cc-engine.el (c-looking-at-or-maybe-in-bracelist): On detection of such a
ref-qualifier, set braceassignp to nil.  When this variable has a nil value,
return nil as the value of the function.  On encountering a } when scanning
backwards, recognise this as the end of a previous construct and stop the
scan.

* decls-47.{cc,res}: New test files.