Optimize for typing characters into long C++ raw strings.
* cc-fonts.el (c-font-lock-complex-decl-prepare, c-font-lock-objc-methods)
(c-font-lock-declarations, c-font-lock-enum-tail)
(c-font-lock-cut-off-declarators, c-font-lock-enclosing-decls): If the chunk
been fontified consists entirely of comments and strings, don't attempt to
perform the function's action.

* cc-mode.el (c-before-change-check-unbalanced-strings): Don't expand
(c-new-BEG c-new-END) unnecessarily to the entire raw string being fontified.
(c-fl-decl-start, c-fl-decl-end): When in a (raw or otherwise) string, don't
return a position outside of the string (which used to cause unneeded
fontification).