CC Mode: Fix multiline block comments in macros.
In particulr, handle multiline block comments whose newlines are not escaped.
There is an example of this in #define EXTRA_CONTEXT_FIELDS in editfns.c.

* cc-engine.el (c-beginning-of-macro, c-end-of-macro): Enclose the loops
scanning escaped newlines with outer loops which check heuristically for,
respectively, a block comment ender and a block comment starter on the lines
we end up on.  (A rigorous syntactic check would be too slow, here.)

* cc-langs.el (c-last-c-comment-end-on-line-re)
(c-last-open-c-comment-start-on-line-re): New language constants/variables.