Fix fontification of first item in CC Mode macro without parentheses
* cc-engine.el (c-find-decl-prefix-search): Handle the new matching
possibility (of a #define construct) in the new c-decl-prefix-or-start-re.
(c-find-decl-spots): Allow the initial search for an in-macro starting point
settle on the # of #define, to facilitate the regexp matching in
c-find-decl-prefix-search.

* cc-langs.el (c-anchored-hash-define-no-parens): New lang const.
(c-literal-start-regexp): Correct what was always supposed to be a "generic
string" regexp element.
(c-decl-prefix-or-start-re): Enhance also to match "#define <identifier>".
(c-dposr-cpp-macro-depth): New lang variable and lang constant

* macro-27.face: Regenerated following correction of fontification.

* macro-31.{c,face}: New test file containing a multiline comment in a macro,
and a struct-like macro without argument parentheses.