Fix problems in CC Mode with " being entered into a comment at EOB.
* cc-engine.el (c-full-lit-near-cache): Amend the definition such that an
element's END element will be nil if the pertinent literal is open at EOB.
(c-full-pp-to-literal): Before setting the aforementioned END element, check
that we're no longer in a literal.  (c-literal-limits): When
c-full-pp-to-literal returns a list with a nil END element, replace this by
(point-max) to keep the interface of c-literal-limits unchanged.

* cc-mode.el (c-after-change-mark-abnormal-strings): Having found a string
quote, check it is not inside an unterminated comment (i.e. one at EOB).