CC Mode.  Fix (c-beginning-of-statement -1) getting stuck with structs.
In particular, with an initialization such as struct foo {..} bar = {...};

* cc-cmds.el (c-forward-to-nth-EOF-{): Rename to c-forward-to-nth-EOF-\;-or-},
and when the starting (or ending) position is in the "variable" part of a
struct/class/union/enum/etc., move to after the terminating semicolon.  Adjust
the counting such that N only gets decremented on a successful movement over
{..}.
(c-beginning-of-defun, c-end-of-defun): Rename the calls to
c-forward-to-nth-EOF-}, as above.