CC Mode.  Allow fontification of "wrong" style comments with warning face.
This fixes bug #4192.

* cc-defs.el (c-font-lock-flush): New macro.

* cc-cmds.el (c-toggle-comment-style): On toggling the comment style, invoke
c-font-lock-flush when c-mark-wrong-style-of-comment is non-nil, to cause that
marking to be done instead on the other style of comment.

* cc-fonts.el (c-maybe-font-lock-wrong-style-comments): New function.
(c-cpp-matchers): Call c-maybe-font-lock-wrong-style-comments when
appropriate.

* cc-vars.el (c-mark-wrong-style-of-comment): New customizable option.

* cc-mode.texi (top level, Indentation Commands, Guessing the Style, Custom
Macros): For some opening quote marks, correct '' to ``.
(Minor Modes): Add an xref to the new page "Wrong Comment Style" in a
footnote.
(Wrong Comment Style): New page.

* comments-8.{cc,face}, comments-9.{cc,face}, comments-10.{awk,face}: New test
files.