CC Mode.  Fix fontification bug with unterminated quotes on adjacent lines
In particular, with these unterminated quotes on each of two adjacent lines,
the following text was spuriously fontified with string face.

* cc-defs.el (c-search-backward-char-property-with-value-on-char): New macro.

* cc-mode.el (c-clear-string-fences): Check whether there is an unmatched
quote at a lower buffer position which should match the current quote, rather
than wrongly assuming the latter is unmatched and marking it with a
punctuation syntax.
(c-font-lock-fontify-region): Ensure all pertinent parts of the buffer have
string fence properties applied before performing any syntactic operations on
it; in particular, this applies to a quote at an earlier buffer position which
"matches" one inside the region about to be fontified.