CC Mode: Protect against consecutive calls to before-change-functions ...
without an intervening call to after-change-functions.  This would have been a
workaround to bug #38691 had the causes of that bug not been removed.

* cc-mode.el (c-just-done-before-change): Add an extra value to this variable,
'whole-buffer, this being set by c-before-change as a signal to c-after-change
that although c-before-change has run, it has assumed the entire buffer as the
change region.
(c-before-change, c-after-change): Adapt to the new meaning of the above.