C++ Mode: handle __attribute__,etc. inside constructor argument lists
This corrects both the fontification and indentation of these things, fixing
bug #42270.

* cc-engine.el (c-do-declarators): Skip over "hangon keys" and noise macros
whilst scanning a putative C++ function.
(c-forward-decl-or-cast-1): When checking for typeless functions, skip over
"hangon keys" and noise macros.

* cc-mode.el (c-fl-decl-end): Deal with certain invalid "nested declarators"
by scanning over them with a recursive call of c-fl-decl-end.

* cc-vars.el (c-noise-macro-names, c-noise-macro-with-parens-names): State in
the doc strings that if either of these is a regexp, it must have a submatch 1
which matches the noise macro exactly.

* constructor-3.{cc,res,face}: New test files.