C++ Mode: Modernize the fontification of "using"
Since "using" is now used in three distinct ways in C++, write a special
function to handle these rather than attempting to adapt the old regular
expressions.

* cc-fonts.el (c-font-lock-declarators): Amend to allow the argument TYPES to
be a face.  This face is given to the declarator being processed.
(c-font-lock-single-decl): Make an argument to c-font-lock-declarators nil or
t, not merely nil or non-nil.
(c-complex-decl-matchers): Include c-font-lock-c++-using in the C++ value of
this variable.
(c-font-lock-c++-using): New function.

* cc-langs.el (c-using-kwds, c-using-key): New lang consts/vars.

* using-2.{cc,face}, using-3.{cc,face}: New test files.