Fix an infinite loop in C++ Mode redisplay.  This was bug #47191.
* cc-defs.el (c-forward-syntactic-ws, c-backward-syntactic-ws): When point is
on the wrong side of a supplied search limit, leave point unmoved rather than
setting it to that limit.

* cc-engine.el (c-forward-name): After scanning a template argument list
(which is not itself subject to a search limit) recalculate the search limit
starting from the end point, since these argument lists can legitimately be
long.  At each of the scanning loops, check point hasn't gone past the limit.