Take account of garbage collection and correct details in previous commit
* cc-fonts.el (c-inhibit-type-finder): New variable.
(c-find-types-background): Check whether we're in a literal, and if so move
forward over it.  Return the final position.
(c-type-finder-timer-func): Rename from c-typeS-.....  Don't perform any
functionality (i) when c-type-finder-time-slot is nil; or (ii) when
font-lock-support-mode is not jit-lock-mode; or (iii) the first time the
function is triggered after a garbage collection.  Use the return value from
c-find-types-background as an optimization.  Several detailed corrections.

* cc-mode.el (c-inhibit-type-finder): New variable.
(c-leave-cc-mode-mode): Handle c-post-gc-hook.
(c-basic-common-init): Initialise c-type-finder-timer to trigger just once,
reinitialising it each time it is called.  Put a function on post-gc-hook.
(c-post-gc-hook): New function.